Flask Web App Setup
This commit comprises of flask file called "app copy.py" to unpack the next uploaded model object called model-1.pkl to generate predictions rendered on an html template present in the next set of upload which is the template folder. Then we have uploaded the requirements.txt that contains core necessities that will be needed in the EC2 instance to run the flask web app. Finally we have the 2 .png files to see the model architecture implemented for this project.